Implemented first MySql commands.
